# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: ploykit API
  version: 0.3.0
  description: |
    ploykit 框架 + example 产品应用的 HTTP 契约。

    认证两通道（webx.Authenticate）：
    - 浏览器会话 Cookie `tk_auth`（写请求须带 X-CSRF-Token，/config 拉取 masked token）；
    - PAT `Authorization: Bearer tk_*`（豁免 CSRF）。

    扁平路由（/api/billing/*、/api/usage、/api/audit*、/api/tasks*）的工作区上下文
    由 `X-Workspace-Id` 请求头注入（@ploykit/client 自动携带）。

    错误信封统一为 ErrorBody（webx.ErrorBody）：`{"error":"E_*","message":string,"details"?}`。
security:
  - sessionCookie: []
  - patBearer: []
...
paths:
...
  /api/notifications:
    get:
      operationId: listNotifications
      tags: [notify]
      summary: 收件箱（按用户隔离，最近优先）
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            default: 20
          description: 条数（非整数或 >100 拒绝 400；≤0 由服务层静默归一为缺省 20——不按违约 400）
      responses:
        '200':
          description: 通知列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notifications/badge:
    get:
      operationId: notificationBadge
      tags: [notify]
      summary: 未读角标数
      responses:
        '200':
          description: 未读数
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadgeResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notify]
      summary: 全部标记已读
      responses:
        '204':
          description: 已全部标记
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notify]
      summary: 单条标记已读
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 通知 ID
      responses:
        '204':
          description: 已标记
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notifications/{id}/archive:
    post:
      operationId: archiveNotification
      tags: [notify]
      summary: 单条归档（收件箱移除、数据保留；重复归档幂等）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 通知 ID
      responses:
        '204':
          description: 已归档
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notification-preferences:
    get:
      operationId: listNotificationPreferences
      tags: [notify]
      summary: 通知偏好（只含显式设置过的行；缺行 = 双渠道默认开启）
      responses:
        '200':
          description: 偏好行列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferenceList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/notification-preferences/{type}:
    put:
      operationId: upsertNotificationPreference
      tags: [notify]
      summary: 更新单类型通知偏好（按用户隔离；至少一个渠道）
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
          description: 通知类型（≤64 字符，见 /api/notification-types 目录）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                email_enabled:
                  type: boolean
                in_app_enabled:
                  type: boolean
      responses:
        '200':
          description: 更新后的偏好行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preference'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/DefaultError'
...
  /api/admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: 站点设置四面（当前值/生效值/默认值/键说明；8 键白名单）
      description: |
        白名单 8 键 = 5 个既有键（banner_text/banner_kind/maintenance_mode/
        allow_signup/allowed_domains）+ 限流三键
        （rate_limit_per_user_per_min / rate_limit_per_ip_per_min /
        rate_limit_allowlist；0 = 关闭限流，
        allowlist 为逗号分隔 IP/CIDR；修改后重启生效，不热重载）。
      responses:
        '200':
          description: 设置全景
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: object, additionalProperties: { type: string } }
                  effective: { type: object, additionalProperties: { type: string } }
                  defaults: { type: object, additionalProperties: { type: string } }
                  descriptions: { type: object, additionalProperties: { type: string } }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/settings/{key}:
    put:
      operationId: adminPutSetting
      tags: [admin]
      summary: 更新站点设置键（白名单外 400）
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        '200': { description: 已更新 }
        '400': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/settings/test-email:
    post:
      operationId: adminTestEmail
      tags: [admin]
      summary: 发送测试邮件（dev 通道回 mode=dev 提示；smtp 回 sent）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, format: email }
      responses:
        '200':
          description: 已发送或 dev 模式提示
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode: { type: string, enum: [dev, smtp] }
                  status: { type: string }
                  hint: { type: string }
        '502': { $ref: '#/components/responses/DefaultError', description: E_MAIL_SEND }
        default: { $ref: '#/components/responses/DefaultError' }
...
  /api/schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: 调度计划列表（schedules:manage；X-Workspace-Id；next_fire_at 正序）
      responses:
        '200':
          description: 计划列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchedulePlan'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: 创建调度计划（schedules:manage；cron 表达式与 IANA 时区经 cronx 校验）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePlanInput'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulePlan'
        '400':
          $ref: '#/components/responses/DefaultError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/schedules/{id}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: 更新计划（启停/换表达式/换时区/换 missed 策略；schedules:manage）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePlanInput'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulePlan'
        '404':
          $ref: '#/components/responses/DefaultError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: 删除计划（schedules:manage；204 无响应体）
      description: 行级删除（非软删）；在途触发由执行端 plan-gone 分支兜底（删除后恰有一发在途 → completed no-op）。重复删除 → 404；跨工作区不可见。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 已删除（无响应体）
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DefaultError'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/schedules/preview:
    post:
      operationId: previewSchedule
      tags: [schedules]
      summary: 预览下次 N 次触发（schedules:manage；cronx 单一事实源，N 默认 3 上限 10）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cron_expr, timezone]
              properties:
                cron_expr: { type: string, description: 5 字段 cron 表达式 }
                timezone: { type: string, description: IANA 时区名（如 Asia/Shanghai） }
                count: { type: integer, minimum: 1, maximum: 10, default: 3 }
      responses:
        '200':
          description: 触发时间列表（UTC ISO8601）
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/DefaultError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tk_auth
      description: 浏览器会话 Cookie（webx.Authenticate 签发/续期；写请求须另带 X-CSRF-Token）
    patBearer:
      type: http
      scheme: bearer
      description: '个人访问令牌（Authorization: Bearer tk_*，CSRF 豁免）'
...
  responses:
    DefaultError:
      description: 其他错误（统一 ErrorBody 信封）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadRequest:
      description: 参数非法（E_VALIDATION / E_BAD_JSON / E_PAYLOAD_TOO_LARGE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthenticated:
      description: 未认证（E_UNAUTHENTICATED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: 无权限（E_FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: 不存在（E_NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Unavailable:
      description: 服务未接线/不可用（E_UNAVAILABLE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 错误码（E_UNAUTHENTICATED / E_FORBIDDEN / E_NOT_FOUND / E_VALIDATION / E_CONFLICT / E_RATE_LIMITED / E_QUOTA_EXCEEDED / E_INTERNAL / E_TIMEOUT / E_PAYLOAD_TOO_LARGE / E_UNAVAILABLE）
        message:
          type: string
        details:
          description: 可选补充（通常为字符串详情）
...
    BadgeResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer

    # ---- identity（字段对齐 packages/ui PKUser） ----
...
    Notification:
      type: object
      required: [id, type, title, body, count, created_at]
      properties:
        id:
          type: string
        type:
          type: string
          description: 通知类型（quota_near_limit / quota_exhausted / ...）
        title:
          type: string
        body:
          type: string
        link:
          type: [string, 'null']
          description: 可选跳转链接
        count:
          type: integer
          description: 合并计数
        read_at:
          type: [string, 'null']
          format: date-time
          description: 已读时间（未读为 null/缺省）
        created_at:
          type: string
          format: date-time
    Preference:
      type: object
      description: 一条显式设置过的通知偏好（缺行 = 双渠道默认开启）
      required: [notification_type, email_enabled, in_app_enabled, updated_at]
      properties:
        notification_type:
          type: string
        email_enabled:
          type: boolean
        in_app_enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
    PreferenceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Preference'
...
    SchedulePlan:
      type: object
      required: [id, workspace_id, kind, cron_expr, timezone, next_fire_at, misfire, created_at, enabled]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
        kind:
          type: string
          description: 执行器类型（如 task.cleanup；未识别 kind 触发时 no-op 记日志）
        cron_expr:
          type: string
          description: 5 字段 cron 表达式（cronx 校验）
        timezone:
          type: string
          description: IANA 时区名
        next_fire_at:
          type: string
          format: date-time
          description: 下次触发（UTC 存储；advance 语义=从当前时间重算，隐式 skip）
        misfire:
          type: string
          enum: [skip, once]
          description: 错过触发策略（默认 skip；once=补发最近一次，60s grace 窗口）
        last_fired_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        enabled:
          type: boolean
    SchedulePlanInput:
      type: object
      required: [kind, cron_expr, timezone]
      properties:
        kind: { type: string }
        cron_expr: { type: string }
        timezone: { type: string }
        misfire:
          type: string
          enum: [skip, once]
        enabled: { type: boolean, default: true }

    # ---- audit ----
...